fix: NoOp encode for web
