+++ dart/lib/src/transport/encode.dart
@@ -1,0 +1,7 @@
+import 'dart:io';
+
+/// Encodes the body using Gzip compression
+void compressBody(List<int> body, Map<String, String> headers) {
+  headers['Content-Encoding'] = 'gzip';
+  body = gzip.encode(body);
+}
