+++ dart/lib/src/transport/noop_encode.dart
@@ -1,0 +1,2 @@
+/// gzip compression is not available on browser
+void compressBody(List<int> body, Map<String, String> headers) {}
